adding payment icon component and code refactoring
