--- app/components/UI/FiatOnRampAggregator/utils/index.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export const fakeAwait = (ms = 1000) => new Promise((resolve) => setTimeout(resolve, ms));
